Source class Place out to seperate file
